add buscar-usuarios

+++ webapp/assets/js/publicacoes.js
@@ -19,7 +19,7 @@ function criarPublicacao(evento) {
     }).done(function() {
         window.location = "/home";
     }).fail(function() {
-        alert("Erro ao criar a publicação!");
+        Swal.fire("Ops...", "Erro ao criar a publicação!", "error");
     })
 }
 
@@ -32,7 +32,7 @@ function curtirPublicacao(evento) {
     elementoClicado.prop('disabled', true);
     $.ajax({
         url: `/publicacoes/${publicacaoId}/curtir`,
-        method: "POST"    
+        method: "POST"
     }).done(function() {
         const contadorDeCurtidas = elementoClicado.next('span');
         const quantidadeDeCurtidas = parseInt(contadorDeCurtidas.text());
@@ -44,7 +44,7 @@ function curtirPublicacao(evento) {
         elementoClicado.removeClass('curtir-publicacao');
 
     }).fail(function() {
-        alert("Erro ao curtir a publicação!");
+        Swal.fire("Ops...", "Erro ao curtir a publicação!", "error");
     }).always(function() {
         elementoClicado.prop('disabled', false);
     });
@@ -59,7 +59,7 @@ function descurtirPublicacao(evento) {
     elementoClicado.prop('disabled', true);
     $.ajax({
         url: `/publicacoes/${publicacaoId}/descurtir`,
-        method: "POST"    
+        method: "POST"
     }).done(function() {
         const contadorDeCurtidas = elementoClicado.next('span');
         const quantidadeDeCurtidas = parseInt(contadorDeCurtidas.text());
@@ -71,28 +71,31 @@ function descurtirPublicacao(evento) {
         elementoClicado.addClass('curtir-publicacao');
 
     }).fail(function() {
-        alert("Erro ao curtir a publicação!");
+        Swal.fire("Ops...", "Erro ao descurtir a publicação!", "error");
     }).always(function() {
         elementoClicado.prop('disabled', false);
     });
 }
 
-function atualizarPublicacao(evento) {
+function atualizarPublicacao() {
     $(this).prop('disabled', true);
 
     const publicacaoId = $(this).data('publicacao-id');
-
+    
     $.ajax({
         url: `/publicacoes/${publicacaoId}`,
         method: "PUT",
         data: {
             titulo: $('#titulo').val(),
             conteudo: $('#conteudo').val()
-        }    
+        }
     }).done(function() {
-        alert("Publicação editada com sucesso!");
+        Swal.fire('Sucesso!', 'Publicação criada com sucesso!', 'success')
+            .then(function() {
+                window.location = "/home";
+            })
     }).fail(function() {
-        alert("Erro ao editar a publicação!");
+        Swal.fire("Ops...", "Erro ao editar a publicação!", "error");
     }).always(function() {
         $('#atualizar-publicacao').prop('disabled', false);
     })
@@ -101,20 +104,31 @@ function atualizarPublicacao(evento) {
 function deletarPublicacao(evento) {
     evento.preventDefault();
 
-    const elementoClicado = $(evento.target);
-    const publicacao = elementoClicado.closest('div')
-    const publicacaoId = publicacao.data('publicacao-id');
+    Swal.fire({
+        title: "Atenção!",
+        text: "Tem certeza que deseja excluir essa publicação? Essa ação é irreversível!",
+        showCancelButton: true,
+        cancelButtonText: "Cancelar",
+        icon: "warning"
+    }).then(function(confirmacao) {
+        if (!confirmacao.value) return;
 
-    elementoClicado.prop('disabled', true);
+        const elementoClicado = $(evento.target);
+        const publicacao = elementoClicado.closest('div')
+        const publicacaoId = publicacao.data('publicacao-id');
     
-    $.ajax({
-        url: `/publicacaoes/${publicacaoId}`,
-        method: "DELETE"
-    }).done(function() {
-        publicacao.fadeOut("slow", function() {
-            $(this).remove();
+        elementoClicado.prop('disabled', true);
+    
+        $.ajax({
+            url: `/publicacoes/${publicacaoId}`,
+            method: "DELETE"
+        }).done(function() {
+            publicacao.fadeOut("slow", function() {
+                $(this).remove();
+            });
+        }).fail(function() {
+            Swal.fire("Ops...", "Erro ao excluir a publicação!", "error");
         });
-    }).fail(function() {
-        alert("Erro ao excluir a publicação!");
-    });
+    })
+
 }
